Move classes ArtificialData, DemoUtils and MultiVariantNormalDistributionRenderer to new sub-package util
