+++ src/test/java/org/mastodon/mamut/fitting/FitEllipsoidPluginDemo.java
@@ -28,6 +28,8 @@
  */
 package org.mastodon.mamut.fitting;
 
+import org.mastodon.mamut.fitting.util.ArtificialData;
+import org.mastodon.mamut.fitting.util.DemoUtils;
 import org.scijava.Context;
