+++ src/test/java/org/mastodon/mamut/fitting/FitEllipsoidPluginTest.java
@@ -35,6 +35,7 @@ import java.util.Arrays;
 import org.junit.Test;
 import org.mastodon.collection.RefObjectMap;
 import org.mastodon.mamut.fitting.ellipsoid.Ellipsoid;
+import org.mastodon.mamut.fitting.util.ArtificialData;
 import org.mastodon.mamut.model.ModelGraph;
 import org.mastodon.mamut.model.Spot;
 import org.scijava.Context;
